+++ src/App.tsx
@@ -1,8 +1,8 @@
 // import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Login from './components/Login/Login';
-import SignUp from './components/SignUp/SignUp';
-import Forgot from './components/Forgot/Forgot';
+import Login from './components/Login';
+import SignUp from './components/SignUp';
+import Forgot from './components/Forgot';
 
 function App() {
   return (
